requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-debug-toolbar==4.3.0
django-storages==1.14.3
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
mysqlclient==2.2.4
packaging==24.0
pillow==10.3.0
pymemcache==4.0.0
python-decouple==3.8
python-dotenv==1.0.1
python-memcached==1.62
requests==2.32.2
sentry-sdk==2.3.1
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
